+++ lib/src/utils/scroll.dart
@@ -3,19 +3,19 @@ import '../_constants.dart';
 @pragma('vm:prefer-inline')
 @pragma('wasm:prefer-inline')
 @pragma('dart2js:prefer-inline')
-String scrollUp([int count = 1]) => '${csi}S' * count;
+String scrollUp([int count = 1]) => count <= 0 ? '' : '$csi${count}S';
 
 @pragma('vm:prefer-inline')
 @pragma('wasm:prefer-inline')
 @pragma('dart2js:prefer-inline')
-String scrollDown([int count = 1]) => '${csi}T' * count;
+String scrollDown([int count = 1]) => count <= 0 ? '' : '$csi${count}T';
 
 @pragma('vm:prefer-inline')
 @pragma('wasm:prefer-inline')
 @pragma('dart2js:prefer-inline')
-String scrollLeft([int count = 1]) => '${csi}Z' * count;
+String scrollLeft([int count = 1]) => count <= 0 ? '' : '$csi$count @';
 
 @pragma('vm:prefer-inline')
 @pragma('wasm:prefer-inline')
 @pragma('dart2js:prefer-inline')
-String scrollRight([int count = 1]) => '${csi}A' * count;
+String scrollRight([int count = 1]) => count <= 0 ? '' : '$csi$count A';